Improved types for packet fields; introduced specific packet types

+++ rest/src/main/java/it/uniud/easyhome/gateway/Gateway.java
@@ -6,7 +6,7 @@ import java.util.Map;
 
 public interface Gateway {
     
-    public int getId();
+    public byte getId();
     
     public ProtocolType getProtocolType();
